Reorganize code into namespaces
