+++ src/interface.cpp
@@ -15,8 +15,11 @@ int main(int argc, char* argv[]) {
         cli_help_section();
         return 0;
     }
-    else if (argc == 2) {   
-        cli_action_detect(argv[1]);
+    else if (argc == 2) {
+        if (argv[1] == "--help" || argv[1] == "-h")
+            cli_help_section();    
+        else
+            cli_action_detect(argv[1]);
     }
     else if (argc >= 3) {
         std::string action = argv[1];
